requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
supabase==2.7.4
python-dotenv==1.0.1
alembic==1.13.2
psycopg2-binary==2.9.9
structlog==24.4.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
cachetools==5.5.0